+++ packages/terra-form-fieldset/tests/wdio/fieldset-spec.js
@@ -10,4 +10,10 @@ Terra.describeViewports('Fieldset', ['tiny', 'medium'], () => {
 
     Terra.validates.element('populated Fieldset with hidden legend');
   });
+
+  it('should display populated Fieldset without legend and helptext', () => {
+    browser.url('/raw/tests/cerner-terra-core-docs/form-fieldset/fieldset-without-legend-and-helptext');
+
+    Terra.validates.accessibility();
+  });
 });