+++ packages/terra-core-docs/src/terra-dev-site/test/form-checkbox/checkbox/DefaultCheckbox.test.jsx
@@ -1,6 +1,13 @@
 import React from 'react';
 import Checkbox from 'terra-form-checkbox';
 
-const checkbox = () => (<Checkbox id="default" labelText="Default Checkbox" />);
+import classNames from 'classnames/bind';
+import styles from './CheckboxTestCommon.module.scss';
 
-export default checkbox;
+const cx = classNames.bind(styles);
+
+export default () => (
+  <div className={cx('spacing')}>
+    <Checkbox id="default" labelText="Default Checkbox" />
+  </div>
+);