+++ packages/terra-core-docs/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Updated
   * Updated jest & wdio tests.
 
+* Added
+  * Added A11y guide and examples for `terra-button`.
+
 ## 1.20.0 - (February 7, 2023)
 
 * Changed