+++ packages/terra-core-docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added 
   * Added Accessibility Guide for `terra-toggle`.
+  * Added A11y guide and examples for `terra-button`.
 
 ## 1.20.0 - (February 7, 2023)